+++ src/Entity/Kubernetes/Cluster/Release.php
@@ -1,0 +1,16 @@
+<?php
+
+namespace Transip\Api\Library\Entity\Kubernetes\Cluster;
+
+class Release extends \Transip\Api\Library\Entity\Kubernetes\Release
+{
+    /**
+     * @var bool
+     */
+    protected $isCompatibleUpgrade;
+
+    public function isCompatibleUpgrade(): bool
+    {
+        return $this->isCompatibleUpgrade;
+    }
+}