+++ src/TransipAPI.php
@@ -55,6 +55,7 @@ use Transip\Api\Library\Repository\Kubernetes\Cluster\NodePools\LabelsRepository
 use Transip\Api\Library\Repository\Kubernetes\Cluster\NodePools\TaintsRepository as KubernetesTaintsRepository;
 use Transip\Api\Library\Repository\Kubernetes\Cluster\NodeRepository as KubernetesNodeRepository;
 use Transip\Api\Library\Repository\Kubernetes\Cluster\Nodes\StatsRepository as KubernetesNodeStatsRepository;
+use Transip\Api\Library\Repository\Kubernetes\Cluster\ReleaseRepository as KubernetesClusterReleaseRepository;
 use Transip\Api\Library\Repository\Kubernetes\ClusterRepository as KubernetesClusterRepository;
 use Transip\Api\Library\Repository\Kubernetes\ProductRepository as KubernetesProductRepository;
 use Transip\Api\Library\Repository\Kubernetes\ReleaseRepository as KubernetesReleaseRepository;
@@ -95,7 +96,7 @@ use Transip\Api\Library\Repository\VpsRepository;
 class TransipAPI
 {
     public const TRANSIP_API_ENDPOINT = "https://api.transip.nl/v6";
-    public const TRANSIP_API_LIBRARY_VERSION = "6.46.0";
+    public const TRANSIP_API_LIBRARY_VERSION = "6.47.0";
     public const TRANSIP_API_DEMO_TOKEN = "eyJ0eXAiOiJKV1QiLCJhbGciOiJSUzI1NiIsImp0aSI6ImN3MiFSbDU2eDNoUnkjelM4YmdOIn0.eyJpc3MiOiJhcGkudHJhbnNpcC5ubCIsImF1ZCI6ImFwaS50cmFuc2lwLm5sIiwianRpIjoiY3cyIVJsNTZ4M2hSeSN6UzhiZ04iLCJpYXQiOjE1ODIyMDE1NTAsIm5iZiI6MTU4MjIwMTU1MCwiZXhwIjoyMTE4NzQ1NTUwLCJjaWQiOiI2MDQ0OSIsInJvIjpmYWxzZSwiZ2siOmZhbHNlLCJrdiI6dHJ1ZX0.fYBWV4O5WPXxGuWG-vcrFWqmRHBm9yp0PHiYh_oAWxWxCaZX2Rf6WJfc13AxEeZ67-lY0TA2kSaOCp0PggBb_MGj73t4cH8gdwDJzANVxkiPL1Saqiw2NgZ3IHASJnisUWNnZp8HnrhLLe5ficvb1D9WOUOItmFC2ZgfGObNhlL2y-AMNLT4X7oNgrNTGm-mespo0jD_qH9dK5_evSzS3K8o03gu6p19jxfsnIh8TIVRvNdluYC2wo4qDl5EW5BEZ8OSuJ121ncOT1oRpzXB0cVZ9e5_UVAEr9X3f26_Eomg52-PjrgcRJ_jPIUYbrlo06KjjX2h0fzMr21ZE023Gw";
 
     /**
@@ -506,6 +507,11 @@ class TransipAPI
     public function kubernetesReleases(): KubernetesReleaseRepository
     {
         return new KubernetesReleaseRepository($this->httpClient);
+    }
+
+    public function kubernetesClusterReleases(): KubernetesClusterReleaseRepository
+    {
+        return new KubernetesClusterReleaseRepository($this->httpClient);
     }
 
     public function openStackProjects(): OpenStackProjectRepository